pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nano-llm-api"
version = "0.1.0"
description = "Minimal local proxy for OpenAI and Anthropic protocol bridging."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "PyYAML>=6.0,<7.0",
  "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0,<9.0"]

[project.scripts]
nano-llm-api = "nano_llm_api.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["nano_llm_api*"]
